Deleted .gitignore from repo
Not something that needs to be on the repo. It's highly system-dependent, not something that needs to be shared.
If I make any changes to it (ex: ignore my 'debug' folder since i don't want to push my system's executable to the repo) it causes problems in pushing/pulling that don't need to happen.